include gridding circles to Gridded class
